Add permutation mutator
