+++ tests/helpers/combiners/test_permutation_combiners.py
@@ -4,14 +4,14 @@ from evol.helpers.combiners.permutation import order_one_crossover, cycle_crosso
 
 
 def test_order_one_crossover_int():
-    seed(45)  # Fix result of select_partition
+    seed(53)  # Fix result of select_partition
     x, y = (8, 4, 7, 3, 6, 2, 5, 1, 9, 0), (0, 1, 2, 3, 4, 5, 6, 7, 8, 9)
     a = order_one_crossover(x, y)
     assert a == (0, 4, 7, 3, 6, 2, 5, 1, 8, 9)
 
 
 def test_order_one_crossover_str():
-    seed(45)  # Fix result of select_partition
+    seed(53)  # Fix result of select_partition
     x, y = ('I', 'E', 'H', 'D', 'G', 'C', 'F', 'B', 'J', 'A'), ('A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J')
     a = order_one_crossover(x, y)
     assert a == ('A', 'E', 'H', 'D', 'G', 'C', 'F', 'B', 'I', 'J')
